WIP: Added common files for models for communication btwn services
