+++ group_service/Dockerfile
@@ -4,11 +4,13 @@ WORKDIR /app
 
 COPY ./requirements.txt /app/requirements.txt
 
+ENV PYTHONPATH=${PYTHONPATH}
+
 RUN apt-get update \
     && apt-get install gcc libpq-dev -y \
     && apt-get clean
 
 RUN pip install -r /app/requirements.txt \
-    && rm -rf /root/.cache/pip
-
+&& rm -rf /root/.cache/pip
+    
 COPY . /app/
