+++ group_service/app/api/models.py
@@ -1,11 +1,1 @@
-from pydantic import BaseModel
-from typing import Optional
-
-class GroupIn(BaseModel):
-    name: str
-
-class GroupOut(GroupIn):
-    id: int
-
-class GroupUpdate(GroupIn):
-    name: Optional[str] = None
+from common.common_models.models import GroupIn, GroupOut, GroupUpdate, StudentOut, StudentIn
