+++ student_service/app/api/models.py
@@ -1,13 +1,1 @@
-from pydantic import BaseModel
-from typing import Optional
-
-class StudentIn(BaseModel):
-    name: str
-    group_id: Optional[int] = None
-
-class StudentOut(StudentIn):
-    id: int
-
-class StudentUpdate(StudentIn):
-    name: Optional[str] = None
-    group_id: Optional[int] = None
+from common.common_models.models import StudentOut, StudentIn, StudentUpdate
